Fix merge from dc09d5 to master

+++ realm/realm-library/src/main/cpp/io_realm_internal_Util.cpp
@@ -29,7 +29,7 @@ using namespace realm::jni_util;
 
 //#define USE_VLD
 #if defined(_MSC_VER) && defined(_DEBUG) && defined(USE_VLD)
-    #include "C:\\Program Files (x86)\\Visual Leak Detector\\include\\vld.h"
+#include "C:\\Program Files (x86)\\Visual Leak Detector\\include\\vld.h"
 #endif
 
 const string TABLE_PREFIX("class_");
@@ -38,25 +38,24 @@ const string TABLE_PREFIX("class_");
 JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void*)
 {
     JNIEnv* env;
-    if (vm->GetEnv((void **) &env, JNI_VERSION_1_6) != JNI_OK) {
+    if (vm->GetEnv((void**)&env, JNI_VERSION_1_6) != JNI_OK) {
         return JNI_ERR;
     }
     else {
         JniUtils::initialize(vm, JNI_VERSION_1_6);
         // Loading classes and constructors for later use - used by box typed fields and a few methods' return value
-        java_lang_long        = GetClass(env, "java/lang/Long");
-        java_lang_long_init   = env->GetMethodID(java_lang_long, "<init>", "(J)V");
-        java_lang_float       = GetClass(env, "java/lang/Float");
-        java_lang_float_init  = env->GetMethodID(java_lang_float, "<init>", "(F)V");
-        java_lang_double      = GetClass(env, "java/lang/Double");
-        java_lang_string      = GetClass(env, "java/lang/String");
+        java_lang_long = GetClass(env, "java/lang/Long");
+        java_lang_long_init = env->GetMethodID(java_lang_long, "<init>", "(J)V");
+        java_lang_float = GetClass(env, "java/lang/Float");
+        java_lang_float_init = env->GetMethodID(java_lang_float, "<init>", "(F)V");
+        java_lang_double = GetClass(env, "java/lang/Double");
+        java_lang_string = GetClass(env, "java/lang/String");
         java_lang_double_init = env->GetMethodID(java_lang_double, "<init>", "(D)V");
-        java_util_date        = GetClass(env, "java/util/Date");
-        java_util_date_init   = env->GetMethodID(java_util_date, "<init>", "(J)V");
+        java_util_date = GetClass(env, "java/util/Date");
+        java_util_date_init = env->GetMethodID(java_util_date, "<init>", "(J)V");
 #if REALM_ENABLE_SYNC
         java_syncmanager = GetClass(env, "io/realm/SyncManager");
 #endif
-
     }
 
     return JNI_VERSION_1_6;
@@ -65,7 +64,7 @@ JNIEXPORT jint JNICALL JNI_OnLoad(JavaVM* vm, void*)
 JNIEXPORT void JNI_OnUnload(JavaVM* vm, void*)
 {
     JNIEnv* env;
-    if (vm->GetEnv((void **) &env, JNI_VERSION_1_6) != JNI_OK) {
+    if (vm->GetEnv((void**)&env, JNI_VERSION_1_6) != JNI_OK) {
         return;
     }
     else {
@@ -83,11 +82,10 @@ JNIEXPORT void JNI_OnUnload(JavaVM* vm, void*)
 
 JNIEXPORT jlong JNICALL Java_io_realm_internal_Util_nativeGetMemUsage(JNIEnv*, jclass)
 {
-    return GetMemUsage();
+    return static_cast<jlong>(GetMemUsage());
 }
 
-JNIEXPORT jstring JNICALL Java_io_realm_internal_Util_nativeGetTablePrefix(
-    JNIEnv* env, jclass)
+JNIEXPORT jstring JNICALL Java_io_realm_internal_Util_nativeGetTablePrefix(JNIEnv* env, jclass)
 {
     realm::StringData sd(TABLE_PREFIX);
     return to_jstring(env, sd);
